Cleaned up some code. Started changing add new league member code to save new league member code so we can also use it edit members.
